pref(user): 优化ddd结构
