+++ internal/user/infra/persistence/repos.go
@@ -2,13 +2,14 @@ package persistence
 
 import (
 	"github.com/cossim/coss-server/internal/user/cache"
-	"github.com/cossim/coss-server/internal/user/domain/user"
+	"github.com/cossim/coss-server/internal/user/domain/repository"
+	"github.com/cossim/coss-server/internal/user/infra/persistence/po"
 	"gorm.io/gorm"
 )
 
 type Repositories struct {
-	UR  user.UserRepository
-	ULR user.UserLoginRepository
+	UR  repository.UserRepository
+	ULR repository.UserLoginRepository
 	db  *gorm.DB
 }
 
@@ -21,5 +22,5 @@ func NewRepositories(db *gorm.DB, cache cache.UserCache) *Repositories {
 }
 
 func (s *Repositories) Automigrate() error {
-	return s.db.AutoMigrate(&UserModel{}, &UserLoginModel{})
+	return s.db.AutoMigrate(&po.User{}, &po.UserLogin{})
 }
